+++ Shiptech.Domain/Exceptions/InvalidD1IIValueException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidD1IIValueException : ShiptechException
     {
-        public InvalidD1IIValueException(int? value) : base(
+        public InvalidD1IIValueException(short? value) : base(
             $"Invalid D1II value: given {value}, required X > 0 and X < 90")
         {
         }
