+++ Shiptech.Domain/Exceptions/InvalidPrefabricationLengthValueException.cs
@@ -4,7 +4,7 @@ namespace Shiptech.Domain.Exceptions
 {
     public class InvalidPrefabricationLengthValueException : ShiptechException
     {
-        public InvalidPrefabricationLengthValueException(int value) : base($"Invalid prefabrication length value: given {value}, required X > 0")
+        public InvalidPrefabricationLengthValueException(short value) : base($"Invalid prefabrication length value: given {value}, required X > 0")
         {
         }
     }
