+++ Shiptech.Domain/ValueObjects/D1II.cs
@@ -4,9 +4,9 @@ namespace Shiptech.Domain.ValueObjects
 {
     public record D1II
     {
-        public int? Value { get; }
+        public short? Value { get; }
 
-        public D1II(int? value)
+        public D1II(short? value)
         {
             if (value is < 0 or > 90)
             {
@@ -16,7 +16,7 @@ namespace Shiptech.Domain.ValueObjects
             Value = value;
         }
 
-        public static implicit operator int?(D1II d1II) => d1II.Value;
-        public static implicit operator D1II(int? d1II) => new(d1II);
+        public static implicit operator short?(D1II d1II) => d1II.Value;
+        public static implicit operator D1II(short? d1II) => new(d1II);
     }
 }
